% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavg_slp.R
\name{get_all_wavg}
\alias{get_all_wavg}
\title{Return 3 weighted averages of slope}
\usage{
get_all_wavg(slp)
}
\arguments{
\item{slp}{A slp object}
}
\value{
a data frame with 3 columns with the output of the 3 methods.
}
\description{
Return 3 weighted averages of slope
}
\examples{
fpath_slp <- system.file("extdata", "071000090603_2.slp", package = "WEPPR")
slp <- read_slp(fpath_slp)
get_all_wavg(slp)
}
