% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavg_slp_sol.R
\name{get_all_wavg_sol}
\alias{get_all_wavg_sol}
\title{Return weighted averages of slope soil object}
\usage{
get_all_wavg_sol(slp, sol)
}
\arguments{
\item{slp}{A slp object}

\item{sol}{A sol object}
}
\value{
a data frame with 3 columns with the output of the 3 methods.
}
\description{
Return weighted averages of slope soil object
}
\examples{
fpath_sol <- system.file("extdata", "071000090603_2.sol", package="WEPPR")
fpath_slp <-system.file("extdata", "071000090603_2.slp", package = "WEPPR")
slp <- read_slp(fpath_slp)
sol <- read_sol(fpath_sol)
get_all_wavg_sol(slp, sol)
}
