% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavg_slp.R
\name{get_wavg}
\alias{get_wavg}
\title{Return weighted average of slope based on weight parameter}
\usage{
get_wavg(slp_extra, weight)
}
\arguments{
\item{slp_extra}{A slp object with additional columns}

\item{weight}{An integer value (0, 1, 2) specifying the method to use for
weighted average calculation}
}
\value{
the weighted average of slope based on the specified weight
}
\description{
Return weighted average of slope based on weight parameter
}
\examples{
fpath_slp <- system.file("extdata", "071000090603_2.slp", package = "WEPPR")
slp <- read_slp(fpath_slp)
slp_extra <- tweak_slp(slp)
get_wavg(slp_extra, weight = 0)
get_wavg(slp_extra, weight = 1)
get_wavg(slp_extra, weight = 2)
}
