% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavg_slp_sol.R
\name{get_wavg_sol}
\alias{get_wavg_sol}
\title{Return weighted average of slp_sol}
\usage{
get_wavg_sol(slp_sol, w_dist, w_thk)
}
\arguments{
\item{slp_sol}{A slp_sol object}

\item{w_dist}{An integer (0, 1, 2) specifying the method to use for
weighted average calculation}

\item{w_thk}{An integer (0, 1, 2) specifying the method to use for
weighted average calculation}
}
\value{
the weighted average of slp_sol
}
\description{
Return weighted average of slp_sol
}
