% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavg_slp_sol.R
\name{getRollAvg}
\alias{getRollAvg}
\title{Return weighted average of x}
\usage{
getRollAvg(x, w)
}
\arguments{
\item{x}{vector of slopes}

\item{w}{vector of distances}
}
\value{
the weighted average of x by w
}
\description{
Return weighted average of x
}
\examples{
x <- c(0, 0.5, 1)
w <- c(1, 2, 3)
getRollAvg(x, w)
}
