% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_cli.R
\name{interpolate_cli}
\alias{interpolate_cli}
\title{Interpolates the climate data frame}
\usage{
interpolate_cli(cli, filename, n = 1001)
}
\arguments{
\item{cli}{a Climate object}

\item{n}{(optional) a positive integer indicating how many equally-spaced time to
calculate.}
}
\value{
a data frame where rows are unique dates and columns are interpolated
ppt values.
}
\description{
Interpolates the climate data frame
}
\examples{
cli <- read_cli(system.file("extdata", "092.63x040.90.cli", package="WEPPR"))
interpolate_cli(cli)

}
\seealso{
\code{\link{pivot_wide_cli_by_date}}
}
