% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_sol.R
\name{interpolate_sol}
\alias{interpolate_sol}
\title{interpolate the soil data file}
\usage{
interpolate_sol(slp_sol)
}
\arguments{
\item{slp_sol}{A Soil object combined with slope data}
}
\value{
a one-row dataframe containing soil data
}
\description{
Converts the soil file into one WEPP run (data frame with one row)
}
\examples{
slp_sol <- merge_slp_sol(slp, sol)
int_sol <- interpolate_sol(slp_sol)

}
