% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_cli.R
\name{pivot_wide_cli_by_date}
\alias{pivot_wide_cli_by_date}
\title{Perfoms a pivot wide on ppt inst values rows based on a single date value}
\usage{
pivot_wide_cli_by_date(cli, n = 1001)
}
\arguments{
\item{cli}{a Climate object}

\item{n}{(optional) a positive integer indicating how many equally-spaced time to
calculate.}
}
\value{
a data frame with only one row and n columns of cumulative ppt
}
\description{
First it performs an expansion on the cli object and the pivot wides the data
}
