% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavg_slp.R
\name{tweak_slp}
\alias{tweak_slp}
\title{Add new columns to slp object for weighted average calculation}
\usage{
tweak_slp(slp)
}
\arguments{
\item{slp}{A slp object}
}
\value{
a data frame with 3 additional distance and slope columns
}
\description{
Creates the following columns:
\itemize{
\item dist_diff (the difference between total distances)
\item avg_slp (rolling average with window size of 2)
\item dist_max (the total distance of each OFE)
}
}
