% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavg_slp.R
\name{wavg}
\alias{wavg}
\title{Return weighted average of slope}
\usage{
wavg(s, d)
}
\arguments{
\item{s}{vector of slopes}

\item{d}{vector of distances}
}
\value{
the weighted average of slope by distance
}
\description{
Return weighted average of slope
}
\examples{
s <- c(0, 0.5, 1)
d <- c(1, 2, 3)
wavg(s, d)
}
