% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavg_slp_sol.R
\name{wavg_solthk}
\alias{wavg_solthk}
\title{Return weighted average of slp_sol by solthk}
\usage{
wavg_solthk(slp_sol)
}
\arguments{
\item{slp_sol}{A slp_sol object}
}
\value{
the weighted average of slp_sol
}
\description{
Return weighted average of slp_sol by solthk
}
